# api/src/api/models/Claim.ts
import { IsNotEmpty } from 'class-validator';

export class Claim {

    public id: string;

    @IsNotEmpty()
    public name: string;

    public toString(): string {
        return `${this.name}`;
    }

}