# api/src/api/swagger.json
{
    "swagger": "2.0",
    "host": "127.0.0.1",
    "info": {
      "title": "",
      "description": "",
      "version": ""
    },
    "basePath": "/",
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "schemes": [],
    "tags": [
      {
        "name": "Search"
      },
      {
        "name": "Registry"
      },
      {
        "name": "Claim"
      }
    ],
    "paths": {
      "/search": {
        "get": {
          "tags": [
            "Search"
          ],
          "summary": "search assets",
          "description": "Returns assets for the search query",
          "parameters": [
            {
              "in": "query",
              "name": "q",
              "type": "string",
              "description": "string used in the search"
            },
            {
              "in": "query",
              "name": "limit",
              "type": "integer",
              "format": "int64",
              "default": 50
            },
            {
              "in": "query",
              "name": "offset",
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "offset": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "limit": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "total": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "data": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Asset"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/registry": {
        "post": {
          "tags": [
            "Registry"
          ],
          "summary": "",
          "description": "",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of asset to return",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Asset object that needs to be added",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewAsset"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "$ref": "#/definitions/Asset"
                  }
                }
              }
            }
          }
        }
      },
      "/registry/{id}": {
        "get": {
          "tags": [
            "Registry"
          ],
          "summary": "",
          "description": "",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of asset to return",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "$ref": "#/definitions/Asset"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Registry"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of asset to update",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Asset object that needs to be updated",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewAsset"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "$ref": "#/definitions/Asset"
                  }
                }
              }
            }
          }
        }
      },
      "/claim": {
        "get": {
          "tags": [
            "Claim"
          ],
          "summary": "",
          "description": "",
          "parameters": [
            {
              "in": "query",
              "name": "limit",
              "type": "integer",
              "format": "int64",
              "default": 50
            },
            {
              "in": "query",
              "name": "offset",
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "offset": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "limit": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "total": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "data": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Claim"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Claim"
          ],
          "summary": "",
          "description": "",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Claim object that needs to be added",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewClaim"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "$ref": "#/definitions/Claim"
                  }
                }
              }
            }
          }
        }
      },
      "/claim/{id}": {
        "get": {
          "tags": [
            "Claim"
          ],
          "summary": "",
          "description": "",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of claim to return",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "$ref": "#/definitions/Claim"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Claim"
          ],
          "summary": "",
          "description": "",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of claim to update",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Claim object that needs to be updated",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewClaim"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "$ref": "#/definitions/Claim"
                  }
                }
              }
            }
          }
        }
      }
    },
    "definitions": {
      "NewAsset": {
        "type": "object",
        "properties": {
            "name": {
                "type":"string"
            },
            "author": {
                "type":"string"
            },
            "description": {
                "type":"string"
            },
            "date": {
                "type":"date-time"
            }
          }
      },
      "NewClaim": {
        "type": "object",
        "properties": {
          "date": {
            "type": "date-time"
          },
          "assetRef":{
              "type" :"object",
              "properties":{
                  "id": {
                    "type":"string"
                  },
                  "version": {
                    "type":"string"
                  }
              }
          },
          "description": {
              "type":"string"
          },
          "history": {
              "type":"array",
              "items": {
                "type":"object",
                "properties":{
                    "date":{
                        "type":"date-time"
                    },
                    "description":{
                        "type":"string"
                    },
                    "userId":{
                        "type":"string"
                    }
                }
              }

          }
        }
      },
      "Asset": {
        "title": "Asset",
        "allOf": [
          {
            "type": "object",
            "properties": {
                "id": {
                "type": "string"
                }
            }
          },
          {
            "$ref": "#/definitions/NewAsset"
          }
        ]
      },
      "Claim": {
        "title": "Claim",
        "allOf": [
          {
            "$ref": "#/definitions/NewClaim"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        ]
      }
    }
  }